Filedelete plugin refactored and fixed. (#441)
* Filedelete plugin refactored and fixed.

* Print about detected deleted or changed files in
filedelete_use_injector mode.
* Some code duplication removed.
* Memory leaks fixed.
* Static counter usage eliminated.
* Don't mix C++ and C style memory management.

* Code Review fixes
